+++ src/components/PageSubtitle.jsx
@@ -1,6 +1,6 @@
 function PageSubtitle({ children }) {
   return (
-    <p className="text-xl md:text-2xl text-blue-600 dark:text-blue-400 mb-8 text-center">{children}</p>
+    <p className="text-body-large md:text-body-large-md text-blue-600 dark:text-blue-400 mb-8 text-center font-body">{children}</p>
   );
 }